+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ThrownExperienceBottleMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
-import insane96mcp.iguanatweaksreborn.module.experience.Experience;
+import insane96mcp.iguanatweaksreborn.module.experience.DroppedExperience;
 import net.minecraft.world.entity.projectile.ThrownExperienceBottle;
 import net.minecraft.world.phys.HitResult;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,6 +13,6 @@ public class ThrownExperienceBottleMixin {
 
 	@Inject(at = @At("HEAD"), method = "onHit")
 	public void onHit(HitResult hitResult, CallbackInfo callbackInfo) {
-		Experience.onXpBottleHit((ThrownExperienceBottle) (Object) this);
+		DroppedExperience.onXpBottleHit((ThrownExperienceBottle) (Object) this);
 	}
 }